+++ src/rh.cpp
@@ -1,7 +1,7 @@
 /*
  * Automatic Watering System
  *
- * (c) 2018 Peter Müller <peter@crycode.de> (https://crycode.de)
+ * (c) 2018-2020 Peter Müller <peter@crycode.de> (https://crycode.de)
  *
  * RadioHead stuff.
  */
@@ -68,10 +68,14 @@ void rhRecv () {
           if (settings.sendAdcValuesThroughRH) {
             rhBufTx[1] |= (1 << 7);
           }
+          // bit 6 indecate if data push is enabled
+          if (settings.pushDataEnabled) {
+            rhBufTx[1] |= (1 << 6);
+          }
           memcpy(&rhBufTx[18], &settings.checkInterval, 2);
-          memcpy(&rhBufTx[20], &settings.dhtInterval, 2);
+          memcpy(&rhBufTx[20], &settings.tempSensorInterval, 2);
 
-          rhSend(RH_MSG_SETTINGS, 22);
+          rhSend(RH_MSG_SETTINGS, 22, RH_FORCE_SEND);
           break;
 
         case RH_MSG_SET_SETTINGS:
@@ -85,12 +89,13 @@ void rhRecv () {
             memcpy(&settings.wateringTime[chan], &rhBufRx[10+chan*2], 2);
           }
           settings.sendAdcValuesThroughRH = ((rhBufRx[1] & (1 << 7)) != 0);
+          settings.pushDataEnabled = ((rhBufRx[1] & (1 << 6)) != 0);
           memcpy(&settings.checkInterval, &rhBufRx[18], 2);
-          memcpy(&settings.dhtInterval, &rhBufRx[20], 2);
+          memcpy(&settings.tempSensorInterval, &rhBufRx[20], 2);
 
           // calc new read times
-          // dht read is 5 seconds before adc read to avoid both readings at the same time
-          dhtNextReadTime = millis() - 5000 + ((uint32_t)settings.dhtInterval * 1000);
+          // temperature sensor read is 5 seconds before adc read to avoid both readings at the same time
+          tempSensorNextReadTime = millis() - 5000 + ((uint32_t)settings.tempSensorInterval * 1000);
           adcNextReadTime = millis() + ((uint32_t)settings.checkInterval * 1000);
           break;
 
@@ -104,22 +109,20 @@ void rhRecv () {
           adcNextReadTime = millis() + 2000;
           break;
 
-        case RH_MSG_TURN_CHANNEL_ON:
-          // turn a channel on
-          if (rhRxLen < 2 || rhBufRx[1] > 3) {
-            return;
-          }
-          // set marker to turn the channel on
-          channelTurnOn[rhBufRx[1]] = true;
-          break;
+        case RH_MSG_TURN_CHANNEL_ON_OFF:
+          if (rhRxLen < 5) return;
 
-        case RH_MSG_TURN_CHANNEL_OFF:
-          // turn a channel on
-          if (rhRxLen < 2 || rhBufRx[1] > 3) {
-            return;
+          for (uint8_t chan = 0; chan < 4; chan++) {
+            if (!settings.channelEnabled[chan]) continue;
+
+            if (rhBufRx[chan + 1] == 0x01 && !channelOn[chan]) {
+              // set marker to turn the channel on
+              channelTurnOn[chan] = true;
+            } else if (rhBufRx[chan + 1] == 0x00 && channelOn[chan]) {
+              // set the turn off time for channel to now
+              channelTurnOffTime[chan] = millis();
+            }
           }
-          // set the turn off time for channel to now
-          channelTurnOffTime[rhBufRx[1]] = millis();
           break;
 
         case RH_MSG_PAUSE:
@@ -132,12 +135,30 @@ void rhRecv () {
           pauseAutomatic = false;
           break;
 
+        case RH_MSG_PAUSE_ON_OFF:
+          // pause on/off
+          if (rhRxLen < 2) {
+            return;
+          }
+          if (rhBufRx[1] == 0x01) {
+            // enable pause
+            pauseAutomatic = true;
+          } else {
+            // resume from pause
+            pauseAutomatic = false;
+          }
+          break;
+
+        case RH_MSG_POLL_DATA:
+          rhSendData(RH_MSG_BATTERY, RH_FORCE_SEND);
+          rhSendData(RH_MSG_CHANNEL_STATE, RH_FORCE_SEND);
+          rhSendData(RH_MSG_TEMP_SENSOR_DATA, RH_FORCE_SEND);
+          rhSendData(RH_MSG_SENSOR_VALUES, RH_FORCE_SEND);
+          break;
+
         case RH_MSG_GET_VERSION:
           // send the software version
-          rhBufTx[1] = SOFTWARE_VERSION_MAJOR;
-          rhBufTx[2] = SOFTWARE_VERSION_MINOR;
-          rhBufTx[3] = SOFTWARE_VERSION_PATCH;
-          rhSend(RH_MSG_VERSION, 4);
+          rhSendData(RH_MSG_VERSION, RH_FORCE_SEND);
           break;
 
         case RH_MSG_PING:
@@ -145,7 +166,7 @@ void rhRecv () {
           for (uint8_t i = 1; i < rhRxLen; i++) {
             rhBufTx[i] = rhBufRx[i];
           }
-          rhSend(RH_MSG_PONG, rhRxLen);
+          rhSend(RH_MSG_PONG, rhRxLen, RH_FORCE_SEND); // use rhSend directly to allow variable data length
           break;
       }
     }
@@ -155,11 +176,12 @@ void rhRecv () {
 /**
  * Function to send a RadioHead message.
  * The data part of the message must be set in rhBufTx before calling this function.
- * @param  msgType Type-code of this message. Will be set in rhBufTx[0].
- * @param  len     Length of the data including the type byte.
- * @return         `true` if the message is successfully send.
+ * @param  msgType   Type-code of this message. Will be set in rhBufTx[0].
+ * @param  len       Length of the data including the type byte.
+ * @param  forceSend Send the message event if push data is disabled. (default false)
+ * @return           `true` if the message is successfully send.
  */
-bool rhSend(uint8_t msgType, uint8_t len, uint8_t delayAfterSend) {
+bool rhSend(uint8_t msgType, uint8_t len, bool forceSend, uint8_t delayAfterSend) {
   rhBufTx[0] = msgType;
   if (!rhManager.sendtoWait(rhBufTx, len, RH_SERVER_ADDR)) {
     blinkCode(BLINK_CODE_RH_SEND_ERROR);
@@ -169,4 +191,91 @@ bool rhSend(uint8_t msgType, uint8_t len, uint8_t delayAfterSend) {
     delay(delayAfterSend);
   }
   return true;
+}
+
+/**
+ * Function to send a RadioHead message with the specified data.
+ * The data part and the length of the message will be automatically set by global variables.
+ * @param  msgType   Type-code of this message. Will be set in rhBufTx[0].
+ * @param  forceSend Send the message event if push data is disabled. (default false)
+ * @return           `true` if the message is successfully send.
+ */
+bool rhSendData(uint8_t msgType, bool forceSend, uint8_t delayAfterSend) {
+  if (!forceSend && !settings.pushDataEnabled) {
+    // to nothing if push data is not enabled and we should not force sending data
+    return true;
+  }
+
+  uint8_t len = 1;
+  switch (msgType) {
+    case RH_MSG_START:
+      // nothing to do
+      break;
+
+    case RH_MSG_CHANNEL_STATE:
+      rhBufTx[1] = channelOn[0] ? 0x01 : 0x00;
+      rhBufTx[2] = channelOn[1] ? 0x01 : 0x00;
+      rhBufTx[3] = channelOn[2] ? 0x01 : 0x00;
+      rhBufTx[4] = channelOn[3] ? 0x01 : 0x00;
+      len = 5;
+      break;
+
+    case RH_MSG_TEMP_SENSOR_DATA:
+      #if TEMP_SENSOR_TYPE == 11 || TEMP_SENSOR_TYPE == 12 || TEMP_SENSOR_TYPE == 22
+        memcpy(&rhBufTx[1], &temperature, 4);
+        memcpy(&rhBufTx[5], &humidity, 4);
+        len = 9;
+      #elif TEMP_SENSOR_TYPE == 1820
+        memcpy(&rhBufTx[1], &temperature, 4);
+        len = 5;
+      #else
+        // nothing to do if no sensor is enabled
+        return true;
+      #endif
+      break;
+
+    case RH_MSG_SENSOR_VALUES:
+      if (!settings.sendAdcValuesThroughRH) {
+        // nothing to do if sending adc values is not enabled
+        return true;
+      }
+
+      for (uint8_t chan = 0; chan < 4; chan++) {
+        if (settings.channelEnabled[chan]) {
+          memcpy(&rhBufTx[1+chan*2], &adcValues[chan], 2);
+        } else {
+          // if channel is disabled but sending adc values is enabled set the value in buffer to 0x0000
+          rhBufTx[1+chan*2] = 0x00;
+          rhBufTx[2+chan*2] = 0x00;
+        }
+      }
+      len = 9;
+      break;
+
+    case RH_MSG_BATTERY:
+      // calc battery percent value
+      if (batteryRaw <= BAT_ADC_LOW) {
+        rhBufTx[1] = 0;
+      } else if (batteryRaw >= BAT_ADC_FULL) {
+        rhBufTx[1] = 100;
+      } else {
+        rhBufTx[1] = 100 * (batteryRaw - BAT_ADC_LOW) / (BAT_ADC_FULL - BAT_ADC_LOW);
+      }
+
+      // store battery raw value into buffer
+      memcpy(&rhBufTx[2], &batteryRaw, 2);
+      len = 4;
+      break;
+
+    case RH_MSG_VERSION:
+        // send the software version
+        rhBufTx[1] = SOFTWARE_VERSION_MAJOR;
+        rhBufTx[2] = SOFTWARE_VERSION_MINOR;
+        rhBufTx[3] = SOFTWARE_VERSION_PATCH;
+        len = 4;
+        break;
+  }
+
+  // send the data
+  return rhSend(msgType, len, forceSend, delayAfterSend);
 }