+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v2.0.0 - 2020-04-13
+- Added support for DS18B20, DS18S20, DS1820, DS1822 temperature sensors
+- Added polling mode
+- Changed some RadioHead messages
+
 ## v1.0.3 - 2018-06-04
 - Added fix for unplausible values read from the DHT sensor
 - Added possibility to change host and port of the control app via env vars